separate savestate from dosbox.cpp
